RJ-38: Change user_id field to optional and refactor find_by_user_id function
The user_id field in the Notification model is changed from String to Option<String> to allow null values. The find_by_user_id function in the notification_repository is refactored for clarity, with an improved comment description and change in the parameter name from 'user_id' to 'id_search' for better understanding. The order of functions is also rearranged for better organization.